% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_rarefaction_curves.R
\name{fast_rarefaction_curves}
\alias{fast_rarefaction_curves}
\title{Rapidly calculate rarefaction curves at specified Hill numbers}
\usage{
fast_rarefaction_curves(physeq, iterations, steps, seed = 1)
}
\arguments{
\item{physeq}{a phyloseq object containing an otu_table of NON-NORMALIZED, INTEGER feature counts. Do not use rarefied, normalized, or relative counts.}

\item{iterations}{how many times alpha diversity should be calculated at each step for each sample}

\item{steps}{how many equally spaced steps at which to rarefy for each sample}

\item{seed}{an integer to set the random seed}
}
\value{
a dataframe with estimated alpha diversity at three hill numbers (hill0 - hill2) at an specified number of equally spaced depths for all samples in long-format.
}
\description{
This function uses the \code{rmvhyper()} function from the \code{extraDistr} package to rapidly estimate rarefaction across your OTU table. This makes it much faster than the \code{rarefy_even_depth} function from \code{phyloseq}, or the manual rarefaction performed in \code{iNEXT}.
}
