% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction_beta_calculations.R
\name{rarefaction_beta_calculations}
\alias{rarefaction_beta_calculations}
\title{Calculate Distance Matrices at Specified Rarefaction Depth and Iteration Count}
\usage{
rarefaction_beta_calculations(
  phy_object,
  rarefy_depth,
  iterations,
  method,
  seed,
  threads
)
}
\arguments{
\item{phy_object}{a phyloseq object containing an otu_table of feature counts}

\item{rarefy_depth}{read depth to rarefy to in each iteration; generally the read count of the sample with the fewest reads}

\item{iterations}{how many times the metric should be calculated before averaging; recommended minimum 100 up to 1000}

\item{method}{a distance metric as listed in the phyloseq::distance() function OR method = "sorensen" (which runs method = 'bray', binary = TRUE under the hood).}

\item{seed}{an integer to set the random seed - required for reproducible calculations! Using set.seed will not be sufficient}

\item{threads}{how many threads to use. Optimum number will depend on your number of iterations. Generally if running 10000 iterations, I wouldn't use more than 20 threads.}
}
\value{
a dist object with average distance between each sample
}
\description{
This function repeatedly rarefies a phyloseq object (usually containing 16S data) and calculates a given distance metric for a set number of iterations. The average distance is then computed for each comparison and the dissimilarity matrix is returned as a \code{dist} object.
}
\examples{
# Calculate Bray-Curtis distance at a rarefied depth of 5000, 100 iterations, using 10 threads
bray_distance <- rarefaction_beta_calculations(phy_object = preprocessed_physeq, rarefy_depth = 5000, iterations = 100, method = "bray", seed = 031491, threads = 10)

}
