% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_rarefaction_curves.R
\name{rarefy_sample}
\alias{rarefy_sample}
\title{Estimate alpha diversity in a given sample of the OTU table}
\usage{
rarefy_sample(otu_vector, steps, iterations, sample_id, seed)
}
\arguments{
\item{otu_vector}{vector of NON-NORMALIZED read counts for each OTU in a single sample}

\item{steps}{how many equally spaced steps to rarefy for each sample}

\item{iterations}{how many times alpha diversity should be calculated at each step for each sample}

\item{sample_id}{ID of sample for which the calculation is being performed}

\item{seed}{an integer to set the random seed}
}
\value{
a dataframe with estimated alpha diversity at three hill numbers (hill0 - hill2) at an specified number of equally spaced depths for that sample.
}
\description{
This function works under the hood in the \code{fast_rarefaction_curves} function; not recommended for user.
}
