% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_flowset_samples.R
\name{rename_flowset_samples}
\alias{rename_flowset_samples}
\title{Interactively Define Gates for Flow Cytometry Flow Set}
\usage{
rename_flowset_samples(flo_set, new_names)
}
\arguments{
\item{flo_set}{A flow set produced using \code{flowCore::read.flowSet()}}
}
\value{
Opens a a shiny app where you can interactively define gates and export them as an .RData object.
}
\description{
This function launches a Shiny app wherein you can interactively define gates for a flow cytometry flow set. Produce your flowset using flowCore::read.flowSet(). Use drop-down menus to select axes for your plots. Click on the plots to draw your gate. Use the Remove Point button to remove vertices, in reverse order that you created them. The Gate Name text input will determine the name of the polygon gate you will create; as such it must follow R guidelines for naming objects. Use Export button to save the gave as an .RData file, which will create a polygonGate as defined by flowCore.
}
