% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_variable_by_clustering.R
\name{reorder_variable_by_clustering}
\alias{reorder_variable_by_clustering}
\title{Reorder categorical variables via hierarchical clustering for plotting heat maps.}
\usage{
reorder_variable_by_clustering(
  data,
  grouping_var1,
  grouping_var2,
  count_var,
  order_both = FALSE,
  value_fill = NULL,
  return_dendro = FALSE
)
}
\arguments{
\item{data}{A dataframe}

\item{grouping_var1}{The first categorical variable you'd like to reorder, as a string (in quotes!)}

\item{grouping_var2}{The second categorical variable you'd like to reorder (you can prevent reordering this variable by setting \code{order_both} to false).}

\item{count_var}{The continuous variable which will fill your heatmap}

\item{order_both}{Should both grouping_var1 and grouping_var2 be reordered? If false (default), only grouping_var1 is reordered.}

\item{value_fill}{Value to fill in for level combinations which don't have observations. Defaults to NULL (no filling).}

\item{return_dendro}{Should the dendrogram also be returned, in addition to the reordered dataframe? If false (the default) returns the dataframe with converted columns. If true, returns a list, in which \code{reord_df} holds the reordered dataframe, and \code{dendrogram} holds the dendrogram. If both grouping variables were reordered, the list will include \code{dendrogram1} for \code{grouping_var1} and \code{dendrogram2} for \code{grouping_var2}.}
}
\value{
Either a dataframe whose grouping columns have been converted to factors whose levels are in the same order as a hierarchical clustering dendrogram, or (if \code{return_dendro = TRUE}) a list holding the reordered dataframe as well as the computed dendrogram(s).
}
\description{
Often in heatmaps, we want to reorder our rows and columns via hierarchical clustering. This function takes in your dataframe and outputs a new dataframe, where your categorical variables have been converted to factors. The levels of these factors are the order of samples in a dendrogram. If wanted, this function can also return the dendrograms, for later plotting.
}
